refactor: restructure file organization in home and tabs modules
